Test in Ubuntu Xenial images with current Python
Current versions of dbus-python have a requirement on dbus >= 1.8, which
isn't met on Ubuntu Trusty. This shouldn't be an issue when deployed, as
the version of dbus-python in use will be whatever is provided by the
distro.

Additionally, this removes testing on Python 3.3, which has been
end-of-lifed, and adds testing on Python 3.7.