Not build for python2 package for Fedora anymore
The bcond conditions don't make any sense to me, I have no idea why
`%bcond_with python2` disabled the python2 ... I would expect the
exact opposite. In fact the whole bcond block seems inverted to me
but I must be misunderstanding it somehow since the rest of the
conditions in the spec is not inverted.

Anyway, I am not rewritting everything but just fixing the build for
F31+.